Moudlarize server.js
