+++ db.json
@@ -4933,10 +4933,10 @@
       "id": 1
     },
     {
-      "email": "Alyce_Morissette11@hotmail.com",
+      "email": "test@test.com",
       "handphone": "09-13363-7054",
       "userid": 2,
-      "password": "moEFZThmvePPdu1",
+      "password": "1234",
       "username": "서현_송",
       "id": 2
     },
